refactor(avatar): Ubi.state 採用 + SettingsPanel 純関数化
- state.ts / systems/AvatarCursorSystem.tsx / systems/AvatarSettingsSystem.tsx /
  systems/utils.ts を削除。共通の cssToState は src/cssToState.ts へ抽出。
- cursor.worker.tsx と settings.worker.tsx それぞれが state + System + 初期化を
  1 ファイルに集約。mutable module state を撤去し Ubi.state.define で宣言的に。
- SettingsPanel.tsx を `renderSettingsPanel(state, actions): VNode` の純関数に。
  state は immutable な snapshot、actions はインターフェース化することで
  snapshot test 可能なシェイプになる。
