+++ plugins/avatar/src/cssToState.ts
@@ -1,3 +1,4 @@
+/** CSS `cursor` プロパティ値を avatar の state キーに正規化する。 */
 export function cssToState(css: string): string {
     if (css === 'pointer') return 'pointer';
     if (css === 'text' || css === 'vertical-text') return 'text';
